+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# Unreleased
+
+### Added
+- exported function `MakeSync` to synchronize database.
+- send custom sync options through the module.
+- send custom sync options through the vuex-offline constructor.
+- added handler function in module to intercept synchronization.
+- added function in constructor to get sync progress.
+- added function in module to get the progress of collection synchronization.
+- possibility to change model to filter in `getFindQuery` helper.
+
+### Fixed
+- filters with the same model was being overwritten.
+
 ## 2.1.0 2021-06-26
 
 ### Added